+++ src/pages/about_page/App.css
@@ -1,0 +1,16 @@
+@import '~antd/dist/antd.css';
+
+.ant-carousel .slick-slide {
+  text-align: center;
+  max-width: 60em;
+  width: 100%;
+  max-height: 30em;
+  height: auto;
+  
+  line-height: 160px;
+  background: #364d79;
+}
+
+.ant-carousel .slick-slide h3 {
+  color: #fff;
+}
